Rename module `see-all-browser` -> `see-all-client`.
All the other client-specific modules use `-client` as a suffix, so this
one should too.
