+++ local-modules/see-all-client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "see-all-browser",
+  "name": "see-all-client",
   "version": "1.0.0",
   "main": "main.js",
